WIP: Caching at store level.
Goal is to cache changes from block updates in-memory until they really
need to be flushed out to sqlite. By doing so, we can batch updates to
sqlite better. This would have the most benefit when the blockchain is
being synced, not actively farming.

Future changes after this could completely separate committing a db
transaction from adding new blocks to the store cache. In this
situation, it would probably be best to have a double buffer for the
store caches (along with a unified interface for accessing the coin and
block stores) so that discarding the in-memory state for what was just
committed would be easy. Handling failures at the database level would
become much more difficult in this case though as newer blocks would
have already been stored in memory (I don't believe there wouldn't
be an error in the consensus protocol, just some state would be lost by
this node if it encountered an error like disk full).